refactor documents pages
